Use proper way to show menu and drawer icons
Because a real overflow menu would show a disgusting drop shadow over
the navigation bar, we can only show a single icon and have to manually
present a bottom sheet with the options.
As for the hamburger icon, it only has a ripple effect (is clickable)
when there is an OnClickListener set.