Use ViewModel, LiveData and Data Binding
It's a two-way Data Binding, so user input automatically updates the
ViewModel and changes in the ViewModel are automatically reflected back
in the UI.
To make it easier, the content layout was moved directly into the
layout of MainActivity. This is not great and should be done using
a Fragment instead.