Migrate content to Fragment
The MainActivity uses a ViewModel to keep track of which Fragment is
currently shown and automatically replace it if the ViewModel is
changed. It seems elegant to just be able to update the ViewModel and
have the fragment transaction happen automatically, but there isn't a
real advantage in terms of code brevity. It's probably simpler to just
run the fragment transaction when a different navigation item is
chosen.