Implement working navigation
Using ViewModel to turn changes in the drawer menu into Fragment
transactions. The conditional instantiation of Fragments (a Fragment is
only replaced if it is not already present) is still a bit of a mess
and working with the menu items and Fragment tags takes way too much
code. There should be a simpler solution than this hell of multiple
switch statements.