Use FAB to insert items into the list
It's a bit stupid that we need to clone the list from the ViewModel
to have DiffUtil see the difference, this inefficiency kind of defeats
the purpose of it.
But since updated data would usually be coming from an external source
in the form of a new list object, this wouldn't be an issue in a real
application. I'm going to leave this in as a showcase for DiffUtil.